+++ database/migrations/2025_02_12_002651_create_admission_test_has_candidate_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('test_id');
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('order_id')->nullable();
-            $table->unsignedTinyInteger('seat_number')->nullable();
+            $table->unsignedSmallInteger('seat_number')->nullable();
             $table->boolean('is_present')->default(false);
             $table->boolean('is_pass')->nullable();
             $table->timestamps();
